Update BLPMain.ox
I've put comments and some alternate code in here.

I don't see where you are trying to call a function and it fails.  What you *might* be confused about is *passing* a function as an argument to another function.  Suppose "solvemyproblem()" is a built-in Ox function that asks you to send it your function:

f() { return 5;}
solvemyproblem(f());
solvemyproblem(f);

There is big difference between those last two statements.  The first one calls f() which returns 5 and then sends that to solvemyproblem:
solvemyproblem(f());  ==> solvemyproblem(5);

The second form sends the function itself, so that solvemyproblem can call it when itself.